[frontend] wip improve overview
