+++ opencti-platform/opencti-front/src/components/Paper.tsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import { PaperProps as MuiPaperProps, Typography, Paper as MuiPaper } from '@mui/material';
+
+interface PaperProps extends MuiPaperProps {
+  title?: string
+}
+
+const Paper = ({ title, children, ...muiProps }: PaperProps) => {
+  return (
+    <div>
+      {title && (
+        <Typography
+          variant="h4"
+          gutterBottom
+          sx={{ display: 'flex', alignItems: 'center', gap: 1 }}
+        >
+          {title}
+        </Typography>
+      )}
+      <MuiPaper
+        variant="outlined"
+        className="paper-for-grid"
+        sx={{
+          marginTop: 1,
+          padding: 2,
+          borderRadius: 0.5,
+          position: 'relative',
+        }}
+        {...muiProps}
+      >
+        {children}
+      </MuiPaper>
+    </div>
+  );
+};
+
+export default Paper;
